fix: do not ignore IPv6 errors in DNSResolveCacheController
Currently, if we try to raise dns listener on IPv6 address and listening fails for any reason, we just silently ignore this for and move on. The reason for this is that raising listener "by default" on IPv6 `::1` sometimes simply doesn't work.
However, if the user decides to use IPv6 service subnet CIDR and `forwardKubeDNSToHost` AND starting dns listening on this
IPv6 network fails for any reason, we will just silently ignore the error, which is not what we want.

After Andrey's (PR)[https://github.com/siderolabs/talos/pull/8583] that will remove `::1` from default `resolv.conf` we no longer have the first problematic part,
so it makes sense to no longer ignore IPv6 dns errors. This commit does this.

Depends on #8583

Signed-off-by: Dmitriy Matrenichev <dmitry.matrenichev@siderolabs.com>