fix: prevent panic in nocloud platform code
The bug was logical: first the check was done for one of the values to
be non-nil, and after that one of the values was assumed to be non-nil,
while it could have been nil.

While fixing that, linter figured out that raw metadata config is never
needed outside of `acquireConfig`, so this got dropped as well,
simplifying the code even more.

Fixes #9578

Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>
(cherry picked from commit 3a0a17ae66dab5c983571fab0f3eac3f87fbc17c)