feat: support link altnames/aliases
At the moment, we don't use/support aliases, but we might in the future.

Altnames are filled out by `systemd-udevd`.

This PR has two parts:

* show aliases & altnames in `LinkStatus`
* match links by aliases/altnames when we configure
  addresses/routes/links

This should make a transition to `systemd-udevd` less painful if the
previous link name is in `altNames`.

Forked rtnetlink for https://github.com/jsimonetti/rtnetlink/pull/241

Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>