fix: disks with 4k sector size and systemd-boot
The previous fix was completely wrong.

See #10097

The core part of the problem was that `dosfstools` was not installed in
the `installer`, so this causes a busybox version of it to be used which
is totally broken for 4k disks.

Adds an integration test.

go-blockdevice updates revert first partition alignment for 4k disks and
allow `talosctl cluster create` to detect properly installation status.

Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>