fix: image cache volume provisioning
There is a logic failure in the `ImageCacheConfig` controller that
`allReady` was always `false` (now should be fixed).

The real issue was that `IMAGECACHE` being in the `SystemDisk` wave, and
not being provisioned until `EPHEMERAL` is not provisioned, which blocks
booting with Image Cache on and without an ISO, by just providing an
empty `IMAGECACHE` disk volume. There isn't much use in this scenario,
but still good to make it fixed.

Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>