fix: ensure proper closure of client.Client.conn with finalizer
As @smira noted in #10263, as soon as compiler can assume that variable is unneeded, it's free to drop it,
and runtime if free to GC it. We want to prevent that, because sometimes variable can be dropped while one
of its fields methods is called or in-flight (the last method in scope for example).

To do this we need to patch two things:
1. Ensure that we call [runtime.KeepAlive](https://pkg.go.dev/runtime#KeepAlive) in all methods after the main
logic path.
2. Replace embedding with unexported fields and add methods (and KeepAlive's) manually. Otherwise, "promoted" methods
do not "keep-alive" our object and GC is free to collect it.

Even tho `ClientConnInterface` interface defines only two methods, we don't know how many type assertions happen
inside, so we export everything from `*grpc.ClientConn` to our type.

Signed-off-by: Dmitriy Matrenichev <dmitry.matrenichev@siderolabs.com>