fix: ignore errors to stop pods
If Talos fails to stop the pods, it doesn't affect any future
operations.

Prevously it was critical for the install/upgrade to succeed, as it
tried to remove whole partition table, but now even dangling mounts
won't affect upgrades.

Sometimes pods can't be stopped because of network file systems.

Fixes #10174

Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>