fix: force DNS runner shutdown on timeout
I observed a random failure (via Sidero Metal integration tests) when
Talos fails to reboot due to the runner never shutting down.

A probable root cause is a bug in the `dns` library (which runs the NDS
server), but we need to ensure this code always terminates, even at the
price of leaking a running DNS server.

Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>