chore: make qemu provider build on darwin
Make the minimal changes required to have the qemu provider build on darwin.
This is the first step in splitting up the darwin qemu provider feature into smaller, easier to review bits.

The following changes will come in the "first crash" order.
Currently the next step that needs fixing is the preflight checks. After that come the various networking bits.

* use a newer commit of the dhcpd fork that has the darwin compatible logic (see https://github.com/insomniacslk/dhcp/pull/550)
* make sure that the (currently broken) darwin qemu logic only runs when `DARWIN_QEMU` env var is set
* avoid using the linux specific cni.ns package (this can be changed back later, just has to be done for now so the thing compiles)
* use a cross-platform fallocate package https://github.com/detailyang/go-fallocate

Signed-off-by: Orzelius <33936483+Orzelius@users.noreply.github.com>
Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>