chore: move the create command to it's own package
This is the first change in the cluster create command refactor project.
The changes will be split into small parts to make review and catching bugs easier.

The command has to be imported in cmd/root.go because it can't be referenced in the mgmt/cluster package as that would create a cyclic import.
This can be fixed later by separating the logic between a third package.

Signed-off-by: Orzelius <33936483+Orzelius@users.noreply.github.com>
Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>