chore: split options between three structs
This is the first step in dividing the provider specific logic from the common operations that need to be performed for all providers.
With this change it's much easier to tell which logic is qemu or docker specific, and which is common.
This also removes the use of package wide variables that are hard to keep track of.

Signed-off-by: Orzelius <33936483+Orzelius@users.noreply.github.com>