docs: fix search on base talos.dev
Because the main talos.dev URL doesn't have a version of any kind in the path (i.e. talos.dev/v1.9/xxx),
the `currentVersion` gets set to "404" when we try to grab the index after splitting.
This PR should work around that, but we should find a better function to use later.

Signed-off-by: Spencer Smith <spencer.smith@talos-systems.com>