fix: improve volume mounter automaton
Fixes #10846

This change would have prevented the seriousness of the issue (riding
into the buffer overrun) by "slowing down" the volume mount operations.

Without this change, an automaton can (depending on the execution flow)
in quick succession lock to the `VolumeMountStatus` which was created
matching previous `VolumeMountRequest`:

* destroy `VolumeMountRequest` - other controller don't have a chance to
  run yet
* automaton restarts
* create `VolumeMountRequest`
* put a finalizer to `VolumeMountStatus`, but this instance of the
  status corresponds to the previously destroyed `VolumeMountRequest`,
  not the new one created

There is no technical problem with the way automaton worked - it just
runs "too quickly" skipping the volume mount/unmount operation, and
simply not letting the volume to be unmounted ever.

With this change, automaton works for its own `VolumeMountStatus` to be
torn down before returning.

Also add unit-tests for the automaton.

Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>