fix: consistently apply dynamic grpc proxy dialer
Use a dialer which picks up environment settings on the fly, which
should allow consistent results when setting environment variables to
control the proxy.

Fixes #10990

There was a race between the first gRPC dial and the moment environment
variables are applied - as gRPC by default reads the environment
variables just once, it might read incomplete state.

Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>