feat: arm64 support for platform vmware
Previous implmentation relied on vmware/vmw-guestinfo, which does not
support ARM64 and is abandoned. Talos-vmtoolsd recently gained an
in-tree implementation of the mechaniscm ("hypercall") to communicate
with the hypervisor. This commit switches Talos to that implementation.

There is one caveat: the library does not have a non-destructive test to
detect if we are a VMWare VM on ARM (it does on x86), and the hypercall
causes a SIGILL when running on a different platforms. In other words,
supplying `talos.platform=vmware` on an ARM non-VM will probably make
Talos die due that SIGILL.

Signed-off-by: Jorik Jonker <jorik.jonker@eu.equinix.com>