fix: userspace wireguard library overrides
Drop using the fork https://github.com/siderolabs/wireguard-go which had
a single change to the variable.

Instead, patch the variable using Go linker flags and only for
`machined`.

This fixes the issue with `wg` utility not working for `talosctl cluster
create --with-siderolink` clsuters, as it was using incompatible path.

We do need to override the path for Talos only, as in Talos `/var/run`
is not always available, but outside of Talos itself we never need this
change at all.

Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>