fix: hold user volume mount point across kubelet restarts
Fixes #11253

Previously the `/var/mnt` was only held by the `kubelet` container, so
on restarts it would retry to unmount it and all child mounts (e.g. user
volumes), and this is not something we actually want.

So instead hold a request for `/var/mnt` always so that kubelet restart
doesn't trigger unmounts.

Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>