fix: talos endpoint might not be created in Kubernetes
This issue is specifically bad with single node controlplanes, and
masked by retries in multiple control planes.

The core issue is that Talos had a race between creating a `talos`
`Service` and `Endpoint`, which might end up with `Endpoint` created
without `Service`, and Kubernetes EndpointController cleans up orphaned
endpoints.

In multiple control planes scenario, and while enabling not on
bootstrap, which might be masked by multiple machines re-creating
endpoints and the issue is not seen.

Fixes #11311

Co-authored-by: Mateusz Urbanek <mateusz.urbanek@siderolabs.com>
Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>