test: fix flaky TestDNS
Fixes #10853

I think the issue is that we do DNS client request to the DNS caching
server, and expect the request to fail on client side, as caching server
doesn't respond in time (as the server is waiting for a response).

Make sure we use a short timeout on our client side, so that we time out
before a server does.

Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>