fix: image cache lockup on a missing volume
When one of the image cache volumes (ISO/disk) is missing,
the controller locks up on shutdown as one of the `VolumeStatus`es is
gone, while other one is locked up on the mount status which is being
held by the finalizer from the controller.

Drop the early exit from the controller to ensure it will keep
processing volumes even if one of them is missing, so that we reach the
code which removes the finalizer.

Fixes #11341

Fixes #11540

Also, unrelated fix for the problem in `talosctl cluster create` with
ISO and bootloader disabled: on reboot QEMU still boots from disk, while
we want to boot from compiled initramfs.

Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>
(cherry picked from commit 92db677b5d32de32ec7e785531b32202e03283b4)