feat(talosctl/cluster): use /etc/resolv.conf nameservers f/ QEMU
`talosctl cluster create` uses the Google and CloudFlare nameservers as
defaults for the QEMU config. This is generally fine, but can cause
issues if these are unreachable from the user's network, and it would
generally be nicer to attempt to use the configured nameservers.

This commit adds logic to read and parse configuration from
`/etc/resolv.conf`. The parsing logic is adapted from the Go stdlib
(see: https://cs.opensource.google/go/go/+/refs/tags/go1.22.1:src/net/dnsconfig_unix.go;l=18).
The current implementation streams the file (I was thinking that in case
something is really wrong and /etc/resolv.conf is way bigger than it
should be, this would be better) but it might be better to keep it
simple and just read the file and then pass it along for parsing.

If anything fails w/ reading/parsing, the previously configured defaults
are used.

Signed-off-by: Laura Brehm <laurabrehm@hey.com>