fix: stop attaching to tearing down mount parents
This bug showed up as a random deadlock on kubelet restart (might be any
other service though).

With a chain of mount requests, like `/var/log` ->
`/var/log/containers`, there was a chance that a new generation of mount
requests might try to pick up a tearing down parent of the previous
generation leading to a deadlock when the mount can't proceed for the
parent.

Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>