fix: drop the Omni API URL check on IP address
I don't see much point in this check, as it's only valuable when joining
to a local development instance of Omni, which is pretty nice usecase.

But this check breaks joining to "real" Omni which has hostname in the
endpoint.

Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>
(cherry picked from commit 0592ff0cdbf54475dc91bfb7c9b9c3047bbe13da)