fix: bond setting change detection
Ignore primaryIndex unless it's both configured and set in the kernel.

I think we should never actually set primaryIndex in any case, but this
fixes an issue with Talos gets into a loop trying to reconcile
primaryIndex (when kernel reports it, and we don't have it set).

This comes from a private user report.

Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>